LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-1_2_1-expected.txt

Inline script attached by DOM manipulation should not run without an 'unsafe-inline' script-src policy, even with default-src *


FAIL Test that securitypolicyviolation event is fired assert_equals: expected "script-src-elem" but got "script-src"
PASS DOM manipulation inline tests
